requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2020.5
sqlparse==0.4.1
wincertstore==0.2
